Start separating validators to two parts
